requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
Flask==3.0.3
gunicorn==23.0.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.6
python-dotenv==1.0.1
